requirements.txt
fastmcp>=0.1.0
pydantic>=2.7.0
pytest>=8.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
moto>=4.2.0
authlib>=1.6.4
pyjwt>=2.8.0
cryptography>=41.0.0
boto3>=1.34.0
